Tests/LibWeb/Text/expected/indexeddb-delete-close-event-ordering.txt

setup.upgradeneeded
setup.success
conn.versionchange
conn.close() called
delete.success
DONE
